builtin: add, equal, not_equal (no tests)

+++ src/common/data_type.rs
@@ -3,7 +3,7 @@
 #[derive(PartialEq, Debug, Clone)]
 pub enum DataType {
     String(String),
-    Int(isize),
+    Int(i32),
     Float(f64),
     Bool(bool),
     // List(Vec<DataType>)
